TripDetailScreen refactor
